+++ lib/github/markup/command_implementation.rb
@@ -1,5 +1,5 @@
 begin
-  require "open3_detach"
+  require "posix-spawn"
 rescue LoadError
   require "open3"
 end
@@ -9,13 +9,17 @@ require "github/markup/implementation"
 
 module GitHub
   module Markup
-    class CommandImplementation < Implementation
-      attr_reader :command, :block
+    class CommandError < RuntimeError
+    end
 
-      def initialize(regexp, command, &block)
+    class CommandImplementation < Implementation
+      attr_reader :command, :block, :name
+
+      def initialize(regexp, command, name, &block)
         super regexp
         @command = command.to_s
         @block = block
+        @name = name
       end
 
       def render(content)
@@ -35,19 +39,34 @@ module GitHub
         end
       end
 
-      def execute(command, target)
-        out = ''
-        Open3.popen3(command) do |stdin, stdout, _|
-          stdin.puts target
-          stdin.close
-          out = stdout.read
+      if defined?(POSIX::Spawn)
+        def execute(command, target)
+          spawn = POSIX::Spawn::Child.new(*command, :input => target)
+          if spawn.status.success?
+            sanitize(spawn.out, target.encoding)
+          else
+            raise CommandError.new(spawn.err.strip)
+          end
         end
-        out.gsub("\r", '')
-      rescue Errno::EPIPE
-        ""
-      rescue Errno::ENOENT
-        ""
+      else
+        def execute(command, target)
+          output = Open3.popen3(*command) do |stdin, stdout, stderr, wait_thr|
+            stdin.puts target
+            stdin.close
+            if wait_thr.value.success?
+              stdout.readlines
+            else
+              raise CommandError.new(stderr.readlines.join('').strip)
+            end
+          end
+          sanitize(output.join(''), target.encoding)
+        end
       end
+
+      def sanitize(input, encoding)
+        input.gsub("\r", '').force_encoding(encoding)
+      end
+
     end
   end
 end